# cache preset info