before:

```
❯ hyperfine "neut check"
Benchmark 1: neut check
  Time (mean ± σ):     626.0 ms ±  18.9 ms    [User: 1014.3 ms, System: 294.5 ms]
  Range (min … max):   601.3 ms … 658.2 ms    10 runs

```

after:

```
❯ hyperfine "neut check"
Benchmark 1: neut check
  Time (mean ± σ):     355.9 ms ±   9.4 ms    [User: 793.5 ms, System: 184.7 ms]
  Range (min … max):   343.4 ms … 369.6 ms    10 runs
```

(`$PWD == (repo-root)/neut/`)